v2(for cloud deployment/requirements.txt
streamlit
httpx
python-dotenv
pandas
pypdf2